+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - always save checkpoint of last epoch model
 - save `best_metrics` value in model checkpoint
 - support for experimental [DETR model](https://github.com/facebookresearch/detr), this model is still unable to be exported caused by the limitation in current exporter design
+- support for per parameter options (see [PyTorch Optimizer](https://pytorch.org/docs/stable/optim.html#per-parameter-options)) using `'param_groups'` key in model components.
+- jit and export global context, see [`vortex/development/networks/modules/utils/config.py`](src/development/vortex/development/networks/modules/utils/config.py)
 - support for per parameter options (see [PyTorch Optimizer](https://pytorch.org/docs/stable/optim.html#per-parameter-options)) 
 using `'param_groups'` key in model components.
 - validation loss calculation to experiment logger
@@ -27,11 +29,13 @@ using `'param_groups'` key in model components.
 - changed `CosineLRScheduler` to `CosineLRWithWarmUp` scheduler
 - changed `TanhLRScheduler` to `TanhLRWithWarmUp` scheduler
 - `BaseTrainer.create_optimizer` only accept model parameters dict (e.g. from `model.parameters()`) instead of the model itself
-
+- update backbone components
+- use mobilenetv3 definition and pretrained from [rwightman](https://github.com/rwightman/pytorch-image-models)
 
 ### Fixed
 - Fix error when using ir_runtime_validate with uneven batch splitting caused by different batch size on the last batch sample
 - Fix error when `save_best_metrics` not present in experiment file
+
 
 
 ## 0.2.0